requirements.txt
pyttsx3
pytest-shutil
pip-custom-platform
py-cpuinfo
subprocess.run
pillow
tkintertable